+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Turn NodeBluePrint into const constructor
 - Make sure disposing a node removes it from the SCM and it's scope
+- SubScopeManager can dynamically create new sub scopes
 
 ### Removed